requirements.txt
Pillow
pdfminer.six
